Separate useful utils into separate helpers file
